week-05-06/weather_cli/requirements.txt
certifi==2026.4.22
charset-normalizer==3.4.7
idna==3.13
python-decouple==3.8
requests==2.33.1
urllib3==2.6.3
